Move test utils to tests/utils.py
