+++ .coveragerc
@@ -1,0 +1,2 @@
+[run]
+source = api
