+++ setup.cfg
@@ -19,3 +19,4 @@ ignore_missing_imports=true
 
 [tool:pytest]
 testpaths=tests
+ignore=tests/utils.py
